requirements.txt
google-genai>=1.0.0
playwright>=1.49.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
httpx>=0.27.0
